[API] Move id specification to domain type
