+++ api/lib/application/memberships/index.js
@@ -2,7 +2,7 @@ const Joi = require('joi');
 
 const securityPreHandlers = require('../security-pre-handlers');
 const membershipController = require('./membership-controller');
-const identifiersType = require('../../domain/validators/id-specification');
+const identifiersType = require('../../domain/types/identifiers-type');
 
 exports.register = async function(server) {
   server.route([
