+++ api/lib/domain/types/identifiers-type.js
@@ -2,7 +2,7 @@ const Joi = require('joi');
 
 const sequenceStartAt = 1;
 const sequencesEndsAt = 2 ** 31 - 1;
-const positive32bitIntegerType = Joi.number().integer().min(sequenceStartAt).max(sequencesEndsAt).required();;
+const positive32bitIntegerType = Joi.number().integer().min(sequenceStartAt).max(sequencesEndsAt).required();
 
 const userId = positive32bitIntegerType;
 const sessionId = positive32bitIntegerType;
